--- back/src/chat/chat.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ChatGateway } from './server';
-
-@Module({
-  providers: [ChatGateway],
-})
-export class ChatModule {}
